requirements.txt
obspy
paho-mqtt
scipy
numpy
scipy
python-dotenv
influxdb-client